requirements_ai_model.txt
streamlit
weaviate-client
langchain
sentence-transformers